requirements.txt
azure-functions
azure-storage-blob
python-dotenv
onnxruntime
numpy
librosa
onnxruntime
soundfile
azure-identity